requirements.txt
transformers>=4.34.0
accelerate>=0.21.0
scipy>=1.11.3
bitsandbytes>=0.41.0
datasets>=2.14.5
sentencepiece>=0.1.99
peft>=0.4.0
deepspeed==0.9.5

evaluate==0.4.0
huggingface-hub==0.16.4
pytorchvideo==0.1.5
rouge-score==0.1.2
tokenizers==0.13.3
torchdata==0.6.1
torchinfo==1.8.0
torchtext==0.15.2
tqdm==4.65.0
nltk==3.8.1
pandas==2.0.3
jsonlines
numpy